# docs/swagger.yaml
basePath: /
definitions:
...
  cicada_web-service-gin_req-res-models.Pool:
    properties:
      Poolli:
        items:
          $ref: '#/definitions/cicada_web-service-gin_req-res-models.Poolli'
        type: array
      created_at:
        type: string
      deleted_at:
        type: string
      gameStatus:
        type: string
      gameid:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  cicada_web-service-gin_req-res-models.Poolli:
    properties:
      Owners:
        items:
          type: string
        type: array
      PoolID:
        type: string
      Value:
        type: number
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreatePoolResponse:
    properties:
      gameID:
        type: string
      poolID:
        type: string
      poollis:
        items:
          $ref: '#/definitions/models.PoolResponse'
        type: array
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.Player:
    properties:
      amount:
        type: number
      name:
        type: string
    type: object
...
  models.Players:
    properties:
      players:
        items:
          $ref: '#/definitions/models.Player'
        type: array
    type: object
...
  models.PoolResponse:
    properties:
      id:
        type: string
      owners:
        items:
          type: string
        type: array
      value:
        type: number
    type: object
  models.PoolToSort:
    properties:
      poolID:
        type: string
      pooliis:
        items:
          $ref: '#/definitions/models.Poollis'
        type: array
    type: object
  models.Poollis:
    properties:
      id:
        type: string
      owners:
        items:
          type: string
        type: array
      poolID:
        type: string
      value:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: cicada REST API description
  title: cicada REST API
  version: "2.0"
paths:
...
  /pool/create-pool:
    post:
      consumes:
      - application/json
      description: Registers pool with their poollis.
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Player data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Players'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreatePoolResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: Registers pool with their poollis.
      tags:
      - Pools
  /pool/delete-pool/{id}:
    delete:
      consumes:
      - application/json
      description: Delete pool with poollis  based on the provided pool ID.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: Delete pool with poollis by pool ID
      tags:
      - Pools
  /pool/get-all-pools:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all pools.
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cicada_web-service-gin_req-res-models.Pool'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: Get all pools with their poolis.
      tags:
      - Pools
  /pool/get-pool/{id}:
    get:
      consumes:
      - application/json
      description: get pool with poollis  based on the provided pool ID.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cicada_web-service-gin_req-res-models.Pool'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: get pool with poollis by pool ID
      tags:
      - Pools
  /pool/get-poollis/{poolID}/{player}:
    get:
      consumes:
      - application/json
      description: get all poollis a player is in  based on the provided player and
        poolID.
      parameters:
      - description: player
        in: path
        name: player
        required: true
        type: string
      - description: poolID
        in: path
        name: poolID
        required: true
        type: string
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cicada_web-service-gin_req-res-models.Poolli'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: get all poollis a player is in
      tags:
      - Pools
  /pool/sort-pool:
    post:
      consumes:
      - application/json
      description: sort pool with number of owners based on the provided poolli Data.
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Player data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PoolToSort'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PoolToSort'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: sort pool with number of owners
      tags:
      - Pools
  /pool/update-pool/{id}:
    put:
      consumes:
      - application/json
      description: updates pool based on the provided id.
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/cicada_web-service-gin_req-res-models.Pool'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cicada_web-service-gin_req-res-models.Pool'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: updates pool by id
      tags:
      - Pools
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"